backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
alembic==1.14.0
bcrypt==4.2.1
google-cloud-speech==2.28.0
